v2/requirements-test.txt
nosetests
websocket-client
smpq
twine
